sRGB space blending that matches other drawing app behavior
It seems like blending is done in the output color space in most drawing applications. Tested in Procreate, Krita, and Clip Studio Paint. This requires somewhat careful handling for blending. During blend operations, sRGB texture format will convert the dst value to linear, and perform blending in linear space, then convert the final blended value back to sRGB for storage. This means we can't just use an sRGB render target for the drawing canvas, and blend normally, as values will be automatically converted to linear space for blending, which isn't what we want. Instead we have to use a non-sRGB render target for the canvas and do the conversion from srgb -> linear manually where required.

Note: There may be a more performant way to do this with an 8 bit canvas (currently we're using a float canvas to avoid precision issues). It might be possible to switch the surface format to non-srgb and simply skip any gamma conversion, since the color should remain sRGB (already gamma corrected) through the entire pipeline.

Currently we're assuming an sRGB output. Later on it would be nice to support wide gaumet and HDR color spaces.